@@ -101,7 +101,7 @@ public class OrderBook
         return trades.ToArray();
     }
 
-    public PriorityQueue<Order> BuyOrders = new PriorityQueue<Order>(1000, new BuyComparer());
-    public PriorityQueue<Order> SellOrders = new PriorityQueue<Order>(1000, new SellComparer());
+    public PriorityQueue<Order> BuyOrders = new (1000, new BuyComparer());
+    public PriorityQueue<Order> SellOrders = new (1000, new SellComparer());
 
 }
